Jobs for new AttackTargeting and new Rotation
Jobs for new AttackTargetting are not implemented yet
New Rotation involves two new jobs and 1 new component. They allow to:
1) Unify all rotations (so there is no need to duplicate logic for rotating to attack target as before)
2) Separately rotate some "Attacking models" of unit (for example turret of tank) (!!!that part is untested as there is no units with such models yet!!!)
Also the logic of rotation was fixed. It is hard to see it visually, but now it is little bit more efficient in some cases + it works predictable (before it was just a luck that it didn't break anywhere, I suppose. Or it was a way how to achieve the similar result with some random code. But still current variant is more efficient in some cases so there is no sense in using previous version at all)